+   Added support for teleport spells that don't randomize the destination location [teleport to fishing guild so you have to have the right fishing level still, or it might teleport you inside the guild] 	-   Added every lunar teleport spell +   Added 'unrepeating messages' 	-   This makes it so your chatbox isn't spammed by things like 'you need to be out of combat for 10 seconds before logging out', or 'a mysterious force stops you from moving'. If you recently saw this message you won't see it again. 	-   Changed default messages from being "unrepeating" +   Added skill restoring tasks 	-   Executed every 100 ticks 	-   Used to bring levels that are too low/too high back to the original level 	-   Rapid restore & berserker prayer actually influence these rates 	-   If you're resting, these rates will be also influenced +   Added health restoration tasks 	-   Brings the health back to the maximum value every 10 ticks by 1. 	-   Added max hitpoints boosts for nex items 		+   On removal of items, if your health is past the maximum health you can possibly have, it is set to your max 			-   Note: this doesn't affect things like hp boosts from brews or rocktails, only nex armour +   Removed all objects from western edgeville room in preparation for dicing. +   Updated leech projectiles slightly, landing graphics id isn't played until the projectile actually lands +   Added blood money into the client +   Added default client settings for newcomers [accept aid = off is important] 	-   Filtering profanity is off, most people are used to it anyway. +   Updated refactored client on git

+++ src/org/redrune/game/node/entity/player/data/PlayerAppearance.java
@@ -236,7 +236,7 @@ public final class PlayerAppearance {
 	 * @param color
 	 * 		The colour.
 	 */
-	public void setColor(byte slot, byte color) {
+	public void setColor(int slot, int color) {
 		this.colors[slot] = color;
 	}
